:+1: Refine and minify `@denops` module
