+++ denops/@denops/mod.ts
@@ -1,2 +1,2 @@
-export { BatchError } from "./denops.ts";
-export type { Context, Denops, Dispatcher, Meta } from "./denops.ts";
+export { BatchError } from "./errors.ts";
+export type { Context, Denops, Dispatcher, Meta } from "./types.ts";
