+++ denops/@denops/test/tester.ts
@@ -2,7 +2,8 @@ import * as path from "https://deno.land/std@0.111.0/path/mod.ts";
 import { Session } from "https://deno.land/x/msgpack_rpc@v3.1.4/mod.ts#^";
 import { using } from "https://deno.land/x/disposable@v1.0.2/mod.ts#^";
 import { deadline } from "https://deno.land/std@0.111.0/async/mod.ts";
-import { Denops, DenopsImpl, Meta } from "../denops.ts";
+import { Denops, Meta } from "../types.ts";
+import { DenopsImpl } from "./impl.ts";
 import { DENOPS_TEST_NVIM, DENOPS_TEST_VIM, run } from "./runner.ts";
 
 const DEFAULT_TIMEOUT = 1000;
